% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demographics_table.R
\name{create_demographics_table}
\alias{create_demographics_table}
\title{Create a standardized table for demographics information}
\usage{
create_demographics_table(data, group = NULL, path = NULL)
}
\arguments{
\item{data}{list of data.frames with SDTM dataset, including at least DM
table.}

\item{group}{name of variable in dataset to group statistics by, e.g.
\code{"ACTARM"}}

\item{path}{optional, path to filename to save output table to.}
}
\description{
Create a standardized table for demographics information
}
