% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nca_table.R
\name{create_nca_table}
\alias{create_nca_table}
\title{Create a table for NCA results}
\usage{
create_nca_table(
  nca_data,
  parameters = NULL,
  group = NULL,
  path = NULL,
  format = "long",
  description = TRUE
)
}
\arguments{
\item{nca_data}{output data from NCA, generated using \code{run_nca()}}

\item{parameters}{vector of parameters to output, using upper case CDISC
standard names, e.g. \code{c("AUCALL", "CMAX", "TMAX")}. Defaults to output all
available parameters.}

\item{group}{name of variable in dataset to group statistics by, e.g.
\code{"ACTARM"}}

\item{path}{optional, path to filename to save output table to.}

\item{format}{output as table in wide or long format.}

\item{description}{include description of parameter names into table?}
}
\description{
Confirms with layout of table commonly used in PK reports, i.e. where
stats are outputted in long format using a single column.
}
