% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_for_modelfit.R
\name{get_data_for_modelfit}
\alias{get_data_for_modelfit}
\title{Get the data needed for model fitting and parse into NONMEM style
format}
\usage{
get_data_for_modelfit(db)
}
\arguments{
\item{db}{list containing database connection details. Requires \code{host},
\code{dbname}, \code{user}, \code{password}, and \code{port}}
}
\value{
NONMEM style dataset as data.frame
}
\description{
Get the data needed for model fitting and parse into NONMEM style
format
}
