% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_for_nca.R
\name{get_data_for_nca}
\alias{get_data_for_nca}
\title{Get the data needed for model fitting and parse into format for easy parsing
with NCA tool.}
\usage{
get_data_for_nca(db = NULL, path = NULL)
}
\arguments{
\item{db}{list with db connection information. If NULL, will try to read
same data from xpt files in \code{path}}

\item{path}{optional. If no \code{db} specified, will read tables from xpt
files in folder \code{path}.}
}
\value{
data.frame
}
\description{
Get the data needed for model fitting and parse into format for easy parsing
with NCA tool.
}
