% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data_for_modelfit.R
\name{parse_data_for_modelfit}
\alias{parse_data_for_modelfit}
\title{Parse data into NONMEM-style dataset}
\usage{
parse_data_for_modelfit(data)
}
\arguments{
\item{data}{list containing SDTM tables as data.frames}
}
\value{
data.frame with population PK input data in NONMEM-style
format. It will also add the non-standard columns ROUTE ("oral", "iv") and
FORM (formulation: "tablet", "suspension", "patch", "infusion", etc.) with
values for each dose and NA for observations.
}
\description{
Parse data into NONMEM-style dataset
}
