% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data_for_nca.R
\name{parse_data_for_nca}
\alias{parse_data_for_nca}
\title{Parse data into dataset for NCA}
\usage{
parse_data_for_nca(data)
}
\arguments{
\item{data}{list containing SDTM tables as data.frames}
}
\value{
data.frame with data for IQNCA package
}
\description{
Parse data into dataset for NCA
}
