% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Get specific tables from database or xpt files}
\usage{
read_data(tables, db = NULL, path = NULL)
}
\arguments{
\item{tables}{Character vector of tables to read}

\item{db}{list with db connection information. If NULL, will try to read
same data from xpt files in \code{path}}

\item{path}{optional. If \code{db} is NULL, will read tables from xpt
files in folder \code{path}.}
}
\value{
List containing each table as a data frame
}
\description{
Get specific tables from database or xpt files
}
