% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data_db}
\alias{read_data_db}
\title{Read in data from a database}
\usage{
read_data_db(tables, conn)
}
\arguments{
\item{tables}{Character vector of tables to load}

\item{conn}{database connection}
}
\value{
List containing each table as a data frame
}
\description{
Read in data from a database
}
\examples{
\dontrun{
# setup
db <- tempfile()
conn <- DBI::dbConnect(RSQLite::SQLite(), db)
DBI::dbWriteTable(conn, "mtcars", mtcars)

dat <- read_data_db(tables = c("mtcars"), conn)

# cleanup
DBI::dbDisconnect(conn)
unlink(db)
}
}
