% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data_files}
\alias{read_data_files}
\title{Read in .xpt files}
\usage{
read_data_files(tables, path = ".")
}
\arguments{
\item{tables}{Character vector of tables to load}

\item{path}{Directory containing .xpt files}
}
\value{
List containing each table as a data frame
}
\description{
Read in .xpt files
}
\examples{
\dontrun{
haven::write_xpt(mtcars, file.path(tempdir(), "mtcars.xpt"))
haven::write_xpt(trees, file.path(tempdir(), "trees.xpt"))
dat <- read_data_files(tables = c("mtcars", "trees"), path = tempdir())
}
}
