% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_modelfit.R
\name{run_modelfit}
\alias{run_modelfit}
\title{Run a modelfit}
\usage{
run_modelfit(data, path, ...)
}
\arguments{
\item{data}{dataset in NONMEM-style format, preferrably created using
\code{create_modelfit_data()}.}

\item{path}{path to file to store output object from fit.}

\item{...}{arguments passed to \code{create_model()} function. See
available arguments to that function.}
}
\description{
This function fits a population PK model to a population PK dataset
that is in a NONMEM-style format. The fit will be performed using
nlmixr2 using the SAEM algorithm. The fit object will be saved to an
rds file, and not be returned.
}
\examples{

run_modelfit(
  data = data,
  n_cmt = 1, 
  route = "oral",
  absorption = "linear",
  path = "./outputs/fit.rds"
)
}
