% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nca.R
\name{run_nca}
\alias{run_nca}
\title{Run a non-compartmental analysis}
\usage{
run_nca(data, path)
}
\arguments{
\item{data}{dataset in pre-parse format, preferrably created using
\code{create_nca_data()}.}

\item{path}{path to file to store output object from fit.}
}
\description{
This function runs an NCA on a pre-parsed dataset. The NCA will be performed
using the IQnca package. The NCA output object will be saved to an
rds file, and not be returned.
}
\examples{
\dontrun{
run_nca(
  data = data,
  path = "./outputs/nca.rds"
)
}
}
