% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sql_query.R
\name{run_sql_query}
\alias{run_sql_query}
\title{Run SQL query against a database}
\usage{
run_sql_query(db, query)
}
\arguments{
\item{db}{list containing database connection details. Requires \code{host},
\code{dbname}, \code{user}, \code{password}, and \code{port}}

\item{query}{SQL query to run}
}
\value{
results from query as data.frame
}
\description{
Based on provided database connection info and query statement.
}
\examples{
db <- list(
  host = "test",
  dbname = "postgres",
  user = "postgres",
  password = "secret",
  port = 5432
)
ex <- run_sql_query(db, "SELECT * FROM ex")

}
