% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_global_plot_theme.R
\name{set_global_plot_theme}
\alias{set_global_plot_theme}
\title{Set global ggplot2 theme and default colors for geoms}
\usage{
set_global_plot_theme(theme = "irx")
}
\arguments{
\item{theme}{theme name}
}
\description{
Set global ggplot2 theme and default colors for geoms
}
