% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data_for_modelfit.R
\name{parse_data_for_modelfit}
\alias{parse_data_for_modelfit}
\title{Parse data into NONMEM-style dataset}
\usage{
parse_data_for_modelfit(data)
}
\arguments{
\item{data}{list containing SDTM tables as data.frames}
}
\value{
data.frame with population PK input data
}
\description{
Parse data into NONMEM-style dataset
}
