% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_modelfit.R
\name{run_modelfit}
\alias{run_modelfit}
\title{Run a modelfit}
\usage{
run_modelfit(data, n_cmt = 1, route = "oral", path)
}
\arguments{
\item{data}{dataset in NONMEM-style format, preferrably created using
\code{create_modelfit_data()}.}

\item{n_cmt}{number of compartments for the population PK model}

\item{route}{administration route, either \code{"oral"} or \code{"iv"}}

\item{path}{path to file to store output object from fit.}
}
\description{
This function fits a population PK model to a population PK dataset
that is in a NONMEM-style format. The fit will be performed using
nlmixr2 using the SAEM algorithm. The fit object will be saved to an
rds file, and not be returned.
}
\examples{

run_modelfit(
  data = data,
  cmt = 1, 
  admin_type = "oral",
  path = "./outputs/fit.rds"
)
}
